requirements.txt
keras==2.7.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
Pillow==9.0.0
plotly==5.5.0
python-dateutil==2.8.2
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
statsmodels==0.13.1
streamlit==1.5.0
tensorflow==2.7.2